backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.7.0
alembic==1.14.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.157.0
python-dateutil==2.9.0
aiosqlite==0.20.0
websockets==14.1
